fill up all stateproof message fields
