+++ stateproof/signer.go
@@ -17,18 +17,13 @@
 package stateproof
 
 import (
-	"errors"
-	"fmt"
 	"time"
 
 	"github.com/algorand/go-algorand/config"
-	"github.com/algorand/go-algorand/crypto"
-	"github.com/algorand/go-algorand/crypto/merklearray"
 	"github.com/algorand/go-algorand/crypto/merklesignature"
 	"github.com/algorand/go-algorand/data/account"
 	"github.com/algorand/go-algorand/data/basics"
 	"github.com/algorand/go-algorand/data/bookkeeping"
-	"github.com/algorand/go-algorand/data/stateproofmsg"
 	"github.com/algorand/go-algorand/protocol"
 )
 
@@ -41,26 +36,6 @@ type sigFromAddr struct {
 	Signer basics.Address            `codec:"signer"`
 	Round  basics.Round              `codec:"rnd"`
 	Sig    merklesignature.Signature `codec:"sig"`
-}
-
-var errInvalidParams = errors.New("provided parameters are invalid")
-var errOutOfBound = errors.New("request pos is out of array bounds")
-
-// The Array implementation for block headers, required to build the merkle tree from them.
-//msgp:ignore
-type blockHeadersArray struct {
-	blockHeaders []bookkeeping.BlockHeader
-}
-
-func (b blockHeadersArray) Length() uint64 {
-	return uint64(len(b.blockHeaders))
-}
-
-func (b blockHeadersArray) Marshal(pos uint64) (crypto.Hashable, error) {
-	if pos >= b.Length() {
-		return nil, fmt.Errorf("%w: pos - %d, array length - %d", errOutOfBound, pos, b.Length())
-	}
-	return b.blockHeaders[pos], nil
 }
 
 func (spw *Worker) signer(latest basics.Round) {
@@ -107,36 +82,6 @@ restart:
 	}
 }
 
-// GenerateStateProofMessage builds a vector commitment from the block headers of the entire interval (up until current round), and returns the root
-// for the account to sign upon. The tree can be stored for performance but does not have to be since it can always be rebuilt from scratch.
-// This is the message that state proofs will attest to.
-func GenerateStateProofMessage(ledger Ledger, latestRoundInInterval basics.Round, stateProofInterval uint64) (stateproofmsg.Message, error) {
-	if latestRoundInInterval < basics.Round(stateProofInterval) {
-		return stateproofmsg.Message{}, fmt.Errorf("GenerateStateProofMessage stateProofRound must be >= than stateproofInterval (%w)", errInvalidParams)
-	}
-	var blkHdrArr blockHeadersArray
-	blkHdrArr.blockHeaders = make([]bookkeeping.BlockHeader, stateProofInterval)
-	firstRound := latestRoundInInterval - basics.Round(stateProofInterval) + 1
-	for i := uint64(0); i < stateProofInterval; i++ {
-		rnd := firstRound + basics.Round(i)
-		hdr, err := ledger.BlockHdr(rnd)
-		if err != nil {
-			return stateproofmsg.Message{}, err
-		}
-		blkHdrArr.blockHeaders[i] = hdr
-	}
-
-	// Build merkle tree from encoded headers
-	tree, err := merklearray.BuildVectorCommitmentTree(blkHdrArr, crypto.HashFactory{HashType: crypto.Sha256})
-	if err != nil {
-		return stateproofmsg.Message{}, err
-	}
-
-	return stateproofmsg.Message{
-		BlockHeadersCommitment: tree.Root().ToSlice(),
-	}, nil
-}
-
 func (spw *Worker) signBlock(hdr bookkeeping.BlockHeader) {
 	proto := config.Consensus[hdr.CurrentProtocol]
 	if proto.StateProofInterval == 0 {
@@ -172,7 +117,7 @@ func (spw *Worker) signBlock(hdr bookkeeping.BlockHeader) {
 	sigs := make([]sigFromAddr, 0, len(keys))
 	ids := make([]account.ParticipationID, 0, len(keys))
 
-	stateproofMessage, err := GenerateStateProofMessage(spw.ledger, hdr.Round, proto.StateProofInterval)
+	stateproofMessage, err := GenerateStateProofMessage(spw.ledger, votersHdr, hdr, proto.StateProofInterval)
 	if err != nil {
 		spw.log.Warnf("spw.signBlock(%d): GenerateStateProofMessage: %v", hdr.Round, err)
 		return
